#11 updates
For loop for creating chain indicator variable works. My issue is with the code below where I'm attempting to assign names in the chain_identity column based on the regex patterns I established above. First two lines work, but the last one for assigning "other" does not. I also tried to change chain_identity based on chain_id, which I thought might be easier, but ran into issues with that as well.